Passing the Playlist Object
Once we have the Playlist information in our Master view, we need to access it in our detail view so that we can see more information about a single playlist. To do this we need to pass the playlist instance to the detail view controller and in this video we take a look at how we can achieve that.